+++ cmd/get/get.go
@@ -6,9 +6,10 @@ import (
 
 var CmdGet = &cobra.Command{
 	Use:   "get",
-	Short: "get info about topics",
+	Short: "get info about topics, consumerGroups",
 }
 
 func init() {
 	CmdGet.AddCommand(cmdGetTopics)
+	CmdGet.AddCommand(cmdGetConsumerGroups)
 }